+++ src/components/DetailCard.jsx
