+++ src/pages/Details.jsx
@@ -2,8 +2,8 @@ import React, { useEffect, useState } from 'react';
 import { Grid, Segment } from 'semantic-ui-react';
 import mockSearchDetails from '../api/searchDetails';
 import PortalLayout from '../layouts/PortalLayout';
-import DetailCard from '../component/DetailCard';
-import PaginationComponent from '../component/Pagination';
+import DetailCard from '../components/DetailCard';
+import PaginationComponent from '../components/Pagination';
 import theme from '../config/theme';
 
 const Details = ({ location: { pathname, search } } = {}) => {
